requirements.txt
## The following requirements were added by pip --freeze:
Babel==1.3
Cython==0.21
DataShape==0.3.0
Flask==0.10.1
Flask-Assets==0.10
Flask-Babel==0.9
Flask-Bootstrap==3.3.0.1
Flask-Environments==0.1
Flask-Hashing==1.1
Flask-Login==0.2.9
Flask-Mail==0.9.1
Flask-Migrate==1.3.0
Flask-OAuth2-Login==0.0.9
Flask-OAuthlib==0.9.1
Flask-OpenID==1.2.4
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF==0.10.3
Flask-WhooshAlchemy==0.56
Flask-Resize==0.8.0
GeoAlchemy2==0.2.4
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
Pillow==2.5.1
PyYAML==3.11
Pygments==1.6
SQLAlchemy==1.1.1
Shapely==1.5.3
Sphinx==1.2.3
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
Whoosh==2.6.0
XlsxWriter==0.5.7
abstract-rendering==0.5.1
alembic==0.7.4
argcomplete==0.8.1
astropy==0.4.2
beautifulsoup4==4.3.2
binstar==0.7.1
bitarray==0.8.1
blaze==0.6.3
blinker==1.3
blz==0.6.2
bokeh==0.6.1
boto==2.32.1
cffi==0.8.6
colorama==0.3.1
configobj==5.0.6
coverage==3.7.1
cryptography==0.5.4
cssmin==0.2.0
cytoolz==0.7.0
decorator==3.4.0
defusedxml==0.4.1
docutils==0.12
factory-boy==2.5.2
flipflop==1.0
future==0.13.1
google-api-python-client==1.4.0
greenlet==0.4.4
guess-language==0.2
gunicorn==19.2.0
h5py==2.3.1
httplib2==0.9
ipython==2.2.0
itsdangerous==0.24
jdcal==1.0
jsmin==2.1.0
llvmpy==0.12.7
lxml==3.4.0
matplotlib==1.4.0
mock==1.0.1
multipledispatch==0.4.7
networkx==1.9.1
nltk==3.0.0
nose==1.3.4
numba==0.14.0
numexpr==2.3.1
numpy==1.9.0
oauth2client==1.4.6
oauthlib==0.7.2
osmapi==0.6.1
openpyxl==1.8.5
pandas==0.14.1
patsy==0.3.0
pbr==0.10.7
ply==3.4
psutil==2.1.1
psycopg2==2.5.4
py==1.4.25
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycosat==0.6.1
pycparser==2.10
pycrypto==2.6.1
pycurl==7.19.5
pyflakes==0.8.1
pyparsing==2.0.1
pytest==2.6.3
python-dateutil==2.1
python3-openid==3.0.5
pytz==2014.7
pyzmq==14.3.1
redis==2.9.1
requests==2.4.1
requests-oauthlib==0.4.2
rope-py3k==0.9.4-1
rsa==3.1.4
runipy==0.1.1
scikit-image==0.10.1
scikit-learn==0.15.2
scipy==0.14.0
simplejson==3.6.5
six==1.8.0
sockjs-tornado==1.0.1
speaklater==1.3
spyder==2.3.1
sqlalchemy-migrate==0.9.2
sqlparse==0.1.14
statsmodels==0.5.0
sympy==0.7.5
tables==3.1.1
toolz==0.7.0
tornado==4.0.2
ujson==1.33
uritemplate==0.6
webassets==0.10.1
xlrd==0.9.3
